Create automateClassification.sh
Need to run this shell script in a `screen` or a `tmux` session to allow it to run in the background continuously. This file automates the image classification part. 
Whenever you get an image from the IOT device, use the deep learning framework ( here Caffe is used ) to classify the image and send the label back.